chore: remove isChanged prop and onChange callbacks (#1915)

+++ src-srv/collaboration/extensions/Repository.ts
@@ -195,7 +195,9 @@ export class RepositoryExtension implements Extension {
       yCtx.delete('isInProgress')
       yCtx.set('hash', createHash(JSON.stringify(doc.getMap('ele').toJSON())))
       yCtx.set('version', result.version)
-      yCtx.delete('isChanged')
+      if (status) {
+        yCtx.delete('isChanged')
+      }
     })
 
     // Cleanup
@@ -238,7 +240,6 @@ export class RepositoryExtension implements Extension {
       const ctx = document.getMap('ctx')
       ctx.set('hash', newHash)
       ctx.set('version', result.version)
-      ctx.delete('isChanged')
     })
 
     return result
